CollectionPage: organize Fixes #1025
